DashBoard updated and divided by components
